+++ NEWS.md
@@ -1,5 +1,11 @@
 # Urgrep News
 
+## v0.2.0 (in progress)
+
+### New features
+- Add support for toggling whether to search in hidden files (`M-s h` in the
+  search prompt, or `urgrep-search-hidden-files` globally)
+
 ## v0.1.1 (2023-06-07)
 
 - Fix Eshell integration